backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Edo Japan Study App — FastAPI backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.6",
    "uvicorn[standard]>=0.34.0",
    "anthropic>=0.43.0",
    "pydantic>=2.10.4",
    "python-dotenv>=1.0.0",
]
